Allow slashes to exist in permalinks
Permalinks in PageViews allow for %variables. The build process allows
for any number of `/` to exist in the path and it'll be normalized. The
serve process is a bit more complicated because of it has a router.

We should allow `/` characters in route parameters so serve can better
match the behavior build has.