Add the ability to automatically set rel and target on external links
* Added three configuration values under the 'links' section:
  - internalhosts: array of hostnames to ignore when deciding to add rel
or target
  - externalrel: value of the rel attribute that should be added to
external links
  - externaltarget: value of the target attribute that should be added
to external links
* For Markdown, inline links are processed based on the configuration
values
* For Twig, an 'a' function was added that takes href, text, and an
array of attribute values and returns an HTML fragment with the 'a' tag.
The configuration values are used to determine if rel or target should
be added to external links. Additionally, it is possible to override the
configuration by passing rel or target as an attribute, and passing an
empty string for either will prevent the attribute from being added at
all.